FIX: remove rendering UX from bookmark model
Fix for `bookmark.js` model. Most logic was moved to `topic` controller
