+++ app/assets/javascripts/discourse/app/models/topic.js
@@ -11,7 +11,6 @@ import Session from "discourse/models/session";
 import Site from "discourse/models/site";
 import User from "discourse/models/user";
 import { ajax } from "discourse/lib/ajax";
-import bootbox from "bootbox";
 import { deepMerge } from "discourse-common/lib/object";
 import discourseComputed from "discourse-common/utils/decorators";
 import { emojiUnescape } from "discourse/lib/text";
@@ -404,73 +403,8 @@ const Topic = RestModel.extend({
     }
   },
 
-  toggleBookmark() {
-    if (this.bookmarking) {
-      return Promise.resolve();
-    }
-    this.set("bookmarking", true);
-    const bookmark = !this.bookmarked;
-    let posts = this.postStream.posts;
-
-    return this.firstPost().then((firstPost) => {
-      const toggleBookmarkOnServer = () => {
-        if (bookmark) {
-          return firstPost.toggleBookmark().then((opts) => {
-            this.set("bookmarking", false);
-            if (opts.closedWithoutSaving) {
-              return;
-            }
-            return this.afterTopicBookmarked(firstPost);
-          });
-        } else {
-          return ajax(`/t/${this.id}/remove_bookmarks`, { type: "PUT" })
-            .then(() => {
-              this.toggleProperty("bookmarked");
-              this.set("bookmark_reminder_at", null);
-              let clearedBookmarkProps = {
-                bookmarked: false,
-                bookmark_id: null,
-                bookmark_name: null,
-                bookmark_reminder_at: null,
-              };
-              if (posts) {
-                const updated = [];
-                posts.forEach((post) => {
-                  if (post.bookmarked) {
-                    post.setProperties(clearedBookmarkProps);
-                    updated.push(post.id);
-                  }
-                });
-                firstPost.setProperties(clearedBookmarkProps);
-                return updated;
-              }
-            })
-            .catch(popupAjaxError)
-            .finally(() => this.set("bookmarking", false));
-        }
-      };
-
-      const unbookmarkedPosts = [];
-      if (!bookmark && posts) {
-        posts.forEach(
-          (post) => post.bookmarked && unbookmarkedPosts.push(post)
-        );
-      }
-
-      return new Promise((resolve) => {
-        if (unbookmarkedPosts.length > 1) {
-          bootbox.confirm(
-            I18n.t("bookmarks.confirm_clear"),
-            I18n.t("no_value"),
-            I18n.t("yes_value"),
-            (confirmed) =>
-              confirmed ? toggleBookmarkOnServer().then(resolve) : resolve()
-          );
-        } else {
-          toggleBookmarkOnServer().then(resolve);
-        }
-      });
-    });
+  deleteBookmark() {
+    return ajax(`/t/${this.id}/remove_bookmarks`, { type: "PUT" });
   },
 
   createGroupInvite(group) {
